+++ boards/adafruit-macropad/examples/adafruit-macropad_blinky.rs
@@ -13,12 +13,15 @@ use adafruit_macropad::{
     },
     Pins, XOSC_CRYSTAL_FREQ,
 };
-use cortex_m_rt::entry;
 use embedded_hal::digital::v2::OutputPin;
 use embedded_time::rate::*;
 use panic_halt as _;
 
-#[entry]
+/// Entry point to our bare-metal application.
+///
+/// The `#[rp2040_hal::entry]` macro ensures the Cortex-M start-up code calls this function
+/// as soon as all global variables and the spinlock are initialised.
+#[rp2040_hal::entry]
 fn main() -> ! {
     let mut pac = pac::Peripherals::take().unwrap();
     let core = pac::CorePeripherals::take().unwrap();