+++ boards/rp-pico/CHANGELOG.md
@@ -14,7 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- None
+- Use `rp2040-hal`'s entry function.
 
 ## 0.4.0 - 2022-06-13